requirements.txt
alembic==1.13.1
alpaca-py==0.14.0
annotated-types==0.6.0
Authlib[flask]
bcrypt==4.1.1
blinker==1.7.0
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.0
cs50==9.3.0
DateTime==5.3
dnspython==2.4.2
docopt==0.6.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-Testing==0.8.1
Flask-WTF==1.2.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
helper==2.5.0
helpers==0.2.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
msgpack==1.0.7
nodejs==0.1.1
numpy==1.26.3
oauthlib==2.1.0
optional-django==0.1.0
packaging==23.2
pandas==2.2.0
password-lib==0.0.3
pipreqs==0.4.13
pluggy==1.3.0
protobuf==4.25.2
py3dns==4.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-helpers==0.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
sseclient-py==1.8.0
template-render==0.1.0
termcolor==2.4.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.4
unzip==1.0.0
uritemplate==4.1.1
urllib3==2.1.0
uuid==1.30
websockets==11.0.3
Werkzeug==3.0.1
WTForms==3.1.1
yarg==0.1.9
zope.interface==6.1
